% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_species_to_genera.R
\name{add_species_to_genera}
\alias{add_species_to_genera}
\title{Adding species to congeneric ones in a phylogenetic tree}
\usage{
add_species_to_genera(specieslist, phy)
}
\arguments{
\item{specieslist}{a community dataset in spacodiR format (see as.spacodi)}

\item{phy}{a phylogenetic tree of class phylo or evolutionary distance matrix between species (see cophenetic.phylo)}
}
\value{
A named list
\itemize{
  \item phylo:  the completed tree 
  \item species_already_present: species already present in the initial tree
  \item non_added_species species not added because of lack of congeneric species in original tree
}
}
\description{
Adding species on a phylogenetic tree when the genus is present
}
\details{
The initial tree must have tip labels in the form of "genus_species")
A list of species to add must be given (specieslist)
Each species is added in turn as sister to a randomly chosen congeneric species
adding a new node with branch length to tips = 0.9 the initial branch length
but with a maximal branch length of 20. => to improve, e.g. considering the distribution
of distance between congeneric spp in the original tree
}
\examples{

}
\author{
Olivier Hardy, Gilles Dauby
}
