% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManipulation.R
\name{as.phylocom}
\alias{as.phylocom}
\title{Converts from spacodi or picante data formats into phylocom format}
\usage{
as.phylocom(data, picante = FALSE, outfile = NULL)
}
\arguments{
\item{data}{species-by-plots matrix}

\item{picante}{logical}

\item{outfile}{an optional text file to which to write output}
}
\value{
A named array, formatted for use in phylocom; note that while the R-object returned by this function has column names, 
if output is written to a file, 
the header is dropped (as appropriate for use in the external phylocom executable: http://www.phylodiversity.net/phylocom/).
}
\description{
Converts from spacodi or picante data formats into phylocom format
}
\details{
This utility converts a species-by-plots matrix into triplet format, 
which is readable by the external program phylocom. If picante=TRUE, 
the data are expected to be in the form used for picante (i.e., a plots-by-species matrix; picante-package).
 If the user selects picante=FALSE, the data are expected to be in the form used for spacodiR (i.e., a species-by-plots matrix). 
 The user has the option to save an output file, defined by outfile.
}
\examples{
# call example data from SPACoDi
data(sp.example)
attach(sp.example)
spl->d.spacodi  
d.spacodi ## SPACoDi format

# convert to phylocom
as.phylocom(data=spl, picante=FALSE)->d.phylocom
d.phylocom ## phylocom format

# convert dataset to picante
as.picante(data=d.phylocom)->d.picante
d.picante ## picante format

# convert back to SPACoDi 
as.spacodi(data=d.picante)


}
\references{
WEBB CO, DD ACKERLY and SW KEMBEL. 2008. Phylocom: software for the analysis of phylogenetic community structure and trait evolution. Bioinformatics 24:2098-2100.
}
\author{
JM Eastman
}
