% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManipulation.R
\name{as.picante}
\alias{as.picante}
\title{Converts from spacodi or phylocom data formats into picante format}
\usage{
as.picante(data, outfile = NULL)
}
\arguments{
\item{data}{a community dataset in phylocom or spacodi format}

\item{outfile}{an optional text file to which to write output}
}
\value{
An array, formatted for use in picante
}
\description{
Converts from spacodi or phylocom data formats into picante format
}
\details{
This utility converts a community dataset (either from phylocom or spacodiR) 
into a format interpretable by picante (see picante-package). phylocom format is also 
referred to as triplet-formatting, where plots are within the first column, abundances 
in the second, and species names in the third column of the dataframe. The user has the 
option to save an output file, defined by outfile. SPACoDi format is similar to that for picante, 
where dataframes between these packages are transposed. SPACoDi format should have species as row names.
}
\examples{
# call example data from SPACoDi
data(sp.example)
attach(sp.example)
spl->d.spacodi  



}
\author{
JM Eastman
}
