% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManipulation.R
\name{as.spacodi}
\alias{as.spacodi}
\title{Cnverts from picante or phylocom data formats into spacodi format.}
\usage{
as.spacodi(data, outfile = NULL)
}
\arguments{
\item{data}{a community dataset in phylocom or spacodi format}

\item{outfile}{an optional text file to which to write output}
}
\value{
An array, formatted for use in spacodiR
}
\description{
Converts from picante or phylocom data formats into spacodi format.
}
\details{
This utility converts a community dataset (either from phylocom or picante (see picante) into a 
format interpretable by either this R-package spacodiR or the external program SPACoDi, a Windows executable
(available at http://ebe.ulb.ac.be/ebe/Software.html). Note also that the community-dataset format used here 
is also that called for by the package vegan; see vegandocs

phylocom format is also referred to as triplet-formatting, where plots are within the first column, 
abundances in the second, and species names in the third column of the dataframe. picante format is 
simply the transpose of spacodiR-formatting of the community dataset: in spacodiR, species are expected 
as the row names of the dataframe, where plots are represented as the column names. The user has the option 
to save an output file, defined by outfile.
}
\examples{
# call example data from spacodiR
data(sp.example)
attach(sp.example)
spl->d.spacodi  
d.spacodi ## SPACoDi format

}
\author{
JM Eastman
}
