% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManipulation.R
\name{check.distmat}
\alias{check.distmat}
\title{Internal function to check matrix structure}
\usage{
check.distmat(obj, tol = 1e-09)
}
\arguments{
\item{obj}{object to check}

\item{tol}{phylo object}
}
\value{
a matrix of phylogenetic distance for all pairs of species
}
\description{
Check matrix structure
}
\author{
JM Eastman
}
