% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cophen.R
\name{cophen}
\alias{cophen}
\title{Internal function that Compute phylogenetic distances among species}
\usage{
cophen(phy)
}
\arguments{
\item{phy}{a phylogenetic tree of class phylo or evolutionary distance matrix between species (see cophenetic.phylo)}
}
\value{
a matrix of phylogenetic distance for all pairs of species
}
\description{
Compute EOO given a data frame of coordinate in decimals degrees
}
\author{
JM Eastman
}
