% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManipulation.R
\name{match.spacodi.data}
\alias{match.spacodi.data}
\title{Internal function that Compute phylogenetic distances among species}
\usage{
match.spacodi.data(
  sp.plot,
  phy = NULL,
  sp.traits = NULL,
  prune = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{sp.plot}{species sample matrix}

\item{phy}{a phylogenetic tree of class phylo}

\item{sp.traits}{species traits matrix}

\item{prune}{logical}

\item{verbose}{logical}
}
\value{
A list of pruned dataset(s)
\itemize{
  \item sp.plot
  \item sp.tree
  \item sp.traits
}
}
\description{
Utility for automating the process of pruning datasets for use in spacodiR
}
\details{
Note that nearly all spacodiR functions require that trees have present (at least) all species sampled in the community dataset
If prune=TRUE, the list of plots pruned will be printed to the console
}
\examples{

# load a species-by-plots matrix, along with a tree
data(sp.example)
attributes(sp.example)
attach(sp.example)
spl
phy
trt

# prune out undersampled plots
spl[,2] <- 0
match.spacodi.data(spl) -> sp.plot.new
sp.plot.new

# match datasets where sp.traits is smaller than the remainder
match.spacodi.data(sp.plot = spl, phy = phy, sp.traits = trt[1:6,])


}
\author{
JM Eastman
}
