% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManipulation.R
\name{prune.sp}
\alias{prune.sp}
\title{Internal function find undersampled plots}
\usage{
prune.sp(sp.plot, verbose = FALSE)
}
\arguments{
\item{sp.plot}{species sample matrix}

\item{verbose}{logical}
}
\value{
a species sample matrix
}
\description{
find undersampled plots
}
\author{
JM Eastman
}
