% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacodi.calc.R
\name{rematrix}
\alias{rematrix}
\title{Internal function}
\usage{
rematrix(arr, names)
}
\arguments{
\item{arr}{vector}

\item{names}{character vector of species names}
}
\value{
a phylo object
}
\description{
Internal function
}
\author{
JM Eastman
}
