% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManipulation.R
\name{reorderspacodiobj}
\alias{reorderspacodiobj}
\title{Internal function find re ordering dataset}
\usage{
reorderspacodiobj(obj, names)
}
\arguments{
\item{obj}{phylo object}

\item{names}{character vector of species names}
}
\value{
a phylo object
}
\description{
re ordering dataset
}
\author{
JM Eastman
}
