% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataResampling.R
\name{resamp.1a}
\alias{resamp.1a}
\title{Rndomizing a community phylogenetics matrix: '1a' of Hardy (2008)}
\usage{
resamp.1a(obj, abund.class.ratio = 4)
}
\arguments{
\item{obj}{a community dataset in spacodiR format (see as.spacodi)}

\item{abund.class.ratio}{numeric vector of length 1 defining abundance classes}
}
\value{
A shuffled dataset
}
\description{
Used for resampling data within a community dataset
}
\details{
A resampling procedure for a species-by-plots matrix, where species are shuffled within abundance classes. 
Species are grouped into distinct abundance classes characterized by a fixed ratio: abund.class.ratio = maximal abundance / minimal abundance. 
For instance, if abund.class.ratio = 4, the limits between abundance classes could be 1, 4, 16, ... . Species are randomly permuted 
within each class, which maintains most of the abundance phylogenetic structure originally present in a dataset.
}
\examples{
data(sp.example)
attach(sp.example)
spl

# shuffle dataset
resamp.1a(obj=spl, abund.class.ratio=3) 

}
\references{
HARDY OJ. 2008. Testing the spatial phylogenetic structure of local communities: statistical performances of different null models and test statistics on a locally neutral community. Journal of Ecology 96:914-926.
}
\author{
Timothy Paine and Jonathan Eastman
}
