% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataResampling.R
\name{resamp.1s}
\alias{resamp.1s}
\title{randomizing a community phylogenetics matrix: '1s' of Hardy (2008)}
\usage{
resamp.1s(obj)
}
\arguments{
\item{obj}{a community dataset in spacodiR format (see as.spacodi)}
}
\value{
A shuffled dataset
}
\description{
Used for resampling data within a community dataset
}
\details{
A resampling procedure for a species-by-plots matrix, where observed abundances are shuffled across species and plots
}
\examples{
data(sp.example)
attach(sp.example)
spl


# shuffle dataset
resamp.1s(obj=spl) 

}
\references{
HARDY OJ. 2008. Testing the spatial phylogenetic structure of local communities: statistical performances of different null models and test statistics on a locally neutral community. Journal of Ecology 96:914-926.
}
\author{
Timothy Paine and Jonathan Eastman
}
