% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataResampling.R
\name{resamp.phy}
\alias{resamp.phy}
\title{partial phylogeny randomization for tips}
\usage{
resamp.phy(phy, node = NULL, time.threshold = 1, proportion = TRUE)
}
\arguments{
\item{phy}{a phylogenetic tree of class phylo}

\item{node}{numeric value(s), specifying the internal node(s) whose tips to shuffle; see nodelabels}

\item{time.threshold}{numeric}

\item{proportion}{logical}
}
\value{
A phylogenetic tree whose tips have been shuffled (without any modification of the underlying topology)
}
\description{
used to shuffle tips subtended by a set of internal nodes determining by divergence time or specified by the user
}
\details{
Either a numeric vector is supplied for node or a time.threshold. If given a set of nodes, 
this function will naively shuffle tips descended from the nodes in the order supplied 
(without regard to whether any internal node in the vector is a descendant of any other node in the node vector). 
If given a time.threshold, tips will be reshuffled within non-nested clades that have a rootmost node that occurs 
within the range [0, time.threshold]. Note that regard to absolute divergence times can be enforced with proportion=FALSE. 
Note further that resamp.phy(phy=phy, node=NULL, node.threshold=1, proportion=TRUE) achieves the same effect as resamp.1s.
}
\examples{
\dontrun{
data(sp.example)
attach(sp.example)

# reshuffle within a time range
time=1/3
bb <- ape::branching.times(phy)
bb <- bb/max(bb)
nodes <- (Ntip(phy)+1):max(phy$edge)
nodes[bb<=time] <- 1
dev.new()
plot(resamp.phy(phy, time.threshold=time, proportion=TRUE))
mtext("reshuffled phylogeny showing affected nodes")
nodelabels(cex=ifelse(nodes==1, 2, NA), col=ifelse(nodes==1, 1, NA), pch=19) 
}

}
\references{
HARDY OJ. 2008. Testing the spatial phylogenetic structure of local communities: statistical performances of different null models and test statistics on a locally neutral community. Journal of Ecology 96:914-926.
}
\author{
Jonathan Eastman
}
