\name{spacodi-internal}
\alias{add.transparency}
\alias{check.distmat}
\alias{cophen}
\alias{debug}
\alias{get.desc.of.node}
\alias{get.descendants.of.node}
\alias{prune.sp}
\alias{reorderspacodiobj}
\alias{spacodi.all.nodes}
\alias{spacodi.exp.nodes}
\alias{withinrange}

\title{internal SPACoDi functions}
\description{\code{spacodi-internal} are not typically called by the user}

\details{
This is an internal \pkg{spacodiR} function, not intended to be called directly by the user.
}