\name{spacodi}

\alias{spacodi}

\docType{package}

\title{main: Spatial and Phylogenetic Analysis of Community Diversity}

\description{
	\pkg{spacodiR} is primarily designed to characterise the structure
	and phylogenetic diversity of communities using abundance or
	presence-absence data of species among community plots. For 
	estimation of community structure, see 
	\code{\link{spacodi.calc}}. More information on \code{SPACoDi} can be found at 
	\url{http://ebe.ulb.ac.be/ebe/Software.html}
}

\details{
 \tabular{ll}{
 Package: \tab spacodi\cr
 Type: \tab Package\cr
 Version: \tab 0.11\cr
 License: \tab GPL\cr
 LazyLoad: \tab yes\cr
 }
}

\author{
Jonathan Eastman, Timothy Paine, and Olivier Hardy <ohardy@ulb.ac.be>
}

\references{
 OLIVIER J. HARDY and BRUNO SENTERRE. 2007. Characterizing the 
  phylogenetic structure of communities by an additive partitioning of 
  phylogenetic diversity. Journal of Ecology 95:493-506.
  
  OLIVIER J. HARDY. 2008. Testing the spatial phylogenetic 
  structure of local communities: statistical performances of 
  different null models and test statistics on a locally neutral 
  community. Journal of Ecology 96:914-926.
}

