% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacodi.calc.R
\name{spacodi.calc}
\alias{spacodi.calc}
\title{Masuring spatial and phylogenetic structuring of diversity in communities}
\usage{
spacodi.calc(
  sp.plot,
  phy = NULL,
  sp.traits = NULL,
  all.together = TRUE,
  prune = TRUE,
  pairwise = FALSE,
  ...
)
}
\arguments{
\item{sp.plot}{a community dataset in spacodiR format (see as.spacodi)}

\item{phy}{a phylogenetic tree of class phylo or evolutionary distance matrix between species (see cophenetic.phylo)}

\item{sp.traits}{a species-by-trait(s) dataframe or a species traits distance matrix (see dist)}

\item{all.together}{Boolean; whether to treat all traits together or separately}

\item{prune}{Boolean whether to dynamically prune datasets if mismatches occur}

\item{pairwise}{Boolean whether to return pairwise diversity measures amongst all plots}

\item{...}{Additional arguments to be passed to match.spacodi.data}
}
\value{
A named list of at least one element (Ist) is returned. The size of the returned list is wholly dependent upon given arguments.

SPECIES DIVERSITY STRUCTURING

\itemize{
  \item Ist: a measure of local species identity excess between individuals, expressing species turnover. 
  It is a form of spatial partition of Gini-Simpson diversity (equivalent to Fst in population genetics). 
  Ist considers only abundances (or presences) in the species-by-plots matrix.
}

PHYLOGENETIC DIVERSITY STRUCTURING

\itemize{
  \item Pst: a measure of local phyletic proximity excess between individuals, expressing species + phylogenetic turnover. 
  It is a form of spatial partition of Rao's quadratic entropy (equivalent to Nst in population genetics). 
  Tst is the analogue for trait data, estimating the spatial partitioning of mean trait-divergence between individuals.
}
}
\description{
Considering species-, phylogenetic-, or trait-diversities, this function measures diversity structuring of community samples.
}
\details{
spacodi.calc requires a community dataset (species-by-plots matrix; sp.plot) of numerical abundance, 
relative abundance, or presence | absence for plots. spacodi.calc returns statistics of diversity partitioning of plots, 
considering species diversity and, if additional information is provided, either trait or phylogenetic diversities among plots. 
If phy=NULL and sp.traits=NULL, a measure of partitioning for species diversity will be returned.
In treating each pair of plots as a community unto its own, pairwise=TRUE will return estimates for diversity structuring 
for all pairwise combinations of plots.

If a phylogeny or trait dataset is supplied with species that are not present in the community dataset (i.e., sp.plot) 
or vice versa, the user has the option to dynamically prune these datasets to match (prune=TRUE). 

If prune=FALSE and dataset mismatches occur, the function will inevitably return NaN where plots have fewer than two distinct species sampled.
}
\examples{
data(sp.example)
attach(sp.example)
spl


# community diversity statistics of Hardy and Senterre (2007): tree-based
spacodi.calc(sp.plot = spl, phy = phy)

# community diversity statistics: trait-based with pairwise comparisons
spacodi.calc(sp.plot = spl, phy = phy, pairwise=TRUE)

# community diversity for a pair of traits
spacodi.calc(sp.plot = spl, sp.traits = trt, all.together=TRUE)


# community diversity for a pair of traits, each singly
spacodi.calc(sp.plot = spl, sp.traits = trt, all.together=FALSE)

# Ist: using abundance data only
spacodi.calc(sp.plot = spl)

# calculations with missing taxa between tree and sp.plot
# excluding the last five species in sp.plot, 
spacodi.calc(sp.plot = spl[1:15,], phy = phy, prune=TRUE)

# as before but with 'manual' pruning of the datasets
match.spacodi.data(sp.plot=spl[1:15,],phy=phy) -> prn.data
spacodi.calc(sp.plot=prn.data$sp.plot, phy=prn.data$sp.tree)
prn.data$sp.plot
prn.data$sp.tree

}
\references{
HARDY OJ and B SENTERRE. 2007. Characterizing the phylogenetic structure of communities by an additive partitioning of phylogenetic diversity. Journal of Ecology 95:493-506.
HARDY OJ. 2008. Testing the spatial phylogenetic structure of local communities: statistical performances of different null models and test statistics on a locally neutral community. Journal of Ecology 96:914-926.
HARDY OJ and L JOST. 2008. Interpreting and estimating measures of community phylogenetic structuring. Journal of Ecology 96:849-852.
}
\author{
Olivier Hardy, Timothy Paine, and Jonathan Eastman
}
