% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_phy.R
\docType{data}
\name{subset_phy}
\alias{subset_phy}
\title{Dataset phylogeny}
\format{
A list with four items
}
\usage{
data(subset_phy)
}
\description{
A phylogeny example
}
\examples{
data(subset_phy)
}
\keyword{subset_phy}
