% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManipulation.R
\name{write.spacodi.data}
\alias{write.spacodi.data}
\title{Generates community phylogenetic datasets}
\usage{
write.spacodi.data(sp.plot, outfile)
}
\arguments{
\item{sp.plot}{a community dataset formatted for the R-package spacodiR}

\item{outfile}{a formatted file for the Windows executable SPACoDi}
}
\value{
a phylo object
}
\description{
Generates community phylogenetic datasets to be used in the external program SPACoDi
}
\details{
This utility writes a species-by-plots matrix into a format readable by the external program SPACoDi, 
a Windows executable
}
\examples{
# generate a community-phylogenetics dataset
data(sp.example)
attach(sp.example)

write.spacodi.data(sp.plot=spl, outfile="spacodi.formatted.txt")

}
\author{
JM Eastman
}
