\name{sp.example}
\alias{sp.example}
\alias{spacodi.data}
\title{example data for spacodiR}
\description{\code{\link[spacodiR]{spacodi-package}} provides example community phylogenetics data, a phylogeny, and trait data}
\usage{data(sp.example)}
\details{This object is a list of example datasets for \pkg{spacodiR}: a community dataset (\code{spl}), 
phylogeny (\code{phy}), species-level traits (\code{trt}), and matrix of distances between plots (\code{dis}). 
The data can be called into R by '\code{data(sp.example)}', where each item is indexed as an element of a list (e.g., for the 
tree, use \code{sp.example$phy}). More generally, documented examples can be run from the console: for instance, try '\code{example(spacodi.calc)}'.}
