% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.example.R
\docType{data}
\name{sp.example}
\alias{sp.example}
\title{Dataset example}
\format{
A list with four items
}
\usage{
data(sp.example)
}
\description{
A list of four elements containing a species-communtiy mmatrix, a phylogeny, a trait-species matrix and community-community distances
}
\examples{
data(sp.example)
}
\keyword{sp.example}
